changed a few variables form int to float

+++ MyProject 4.18/Source/MyProject/MyProjectCharacter.h
@@ -137,9 +137,9 @@ public:								////// PUBLIC //////
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = GeneralMovementCPP)
 		int walkSpeed = 600;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Crank)
-		int notMoving = 5;
+		float notMoving = 5;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Gameplay)
-		int losingHealthTimer = 1;
+		float losingHealthTimer = 1;
 
 
 	// is Slomo Active or Deactive
